got some basic abstraction done. Will need review
