+++ index.html
@@ -11,7 +11,7 @@
       <th>Calories</th>
     </tr>
   </table>
-  <label class="food-filter">Filter by Name:
+  <label class="foods-filter">Filter by Name:
     <input class="search-value">
   </label>
