+++ lib/foods.js
@@ -2,158 +2,40 @@
 // of specific herokuUrl api request enpoint
 // make requests with `herokuUrl + "api/v1/foods"` etc
 const herokuUrl = require("./heroku-url")
+const CruddyFood = require("./cruddy-food")
+const HtmlEvents = require("./html-events")
+const Filter = require("./filter")
 const $ = require("jQuery")
 
-$('#newFoodForm').hide()
+class DefaultFoodLoader {
+  constructor() {
+    this.crud  = new CruddyFood()
+    this.html  = new HtmlEvents()
+    this.filter = new Filter()
+  }
 
-function getFoods() {
-  $.ajax({
-    type: "GET",
-    url: `${herokuUrl()}api/v1/foods`,
-    success: function(posts) {
-      launchPage(posts)
-    }
-  })
-}
-
-function launchPage(posts) {
-  fillTable(posts)
-  deleteFoodListener()
-  editFoodListener()
-  foodFilterListenter()
-  createFoodButtonListener()
-}
-
-function deleteFoodListener() {
-$('.deleteButton').on('click', function() {
-  $.ajax({
-    type: 'DELETE',
-    url: herokuUrl() + `api/v1/foods/${event.target.id}`,
-    success: $(`#${event.target.id}`).hide()
-    })
-  })
-}
-
-function editFoodListener(){
-  $('.foodsTable').on('keydown', '.food', function(event) {
-    const foodRow  = $(this).parent()
-    const foodId   = foodRow.attr("id")
-    const attrName = $(this).attr("name")
-    const data     = { [attrName]: $(this).text() }
-
-    if (event.keyCode === 13) {
-      ajaxPatchRequest(data, foodId)
-      return false
-    }
-  })
-}
-
-function ajaxPatchRequest(data, foodId) {
-  $.ajax({
-    type: 'PATCH',
-    url: `${herokuUrl()}api/v1/foods/${foodId}`,
-    data: data,
-    success: function() {
-      alert("Successful")
-    },
-    error: function() {
-      alert("Nope")
-    }
-  })
-}
-
-function foodFilterListenter(){
-  $(`.food-filter`).on('keydown', function(event) {
-    let searchValue = $('.search-value').val().toLowerCase()
+  getFoods() {
+    $('#newFoodForm').hide()
+    let page = this
     $.ajax({
-      type: 'GET',
+      type: "GET",
       url: `${herokuUrl()}api/v1/foods`,
       success: function(posts) {
-        appendFood(posts, searchValue)
+        page.launchPage(posts)
       }
     })
-  })
+  }
+
+  launchPage(posts) {
+    this.html.fillTable(posts)
+    this.filter.startListener("foods", this.filter)
+    this.crud.startListener()
+  }
 }
 
-function appendFood(foods, searchValue) {
-   let foodObjects = foods.filter(function(food) {
-    return food.name.toLowerCase().includes(searchValue)
-  })
+$(document).ready(function() {
+  const loadPage = new DefaultFoodLoader()
+  loadPage.getFoods()
+})
 
-  $('.food').remove()
-  $('.deleteButton').hide()
-
-
-  fillTable(foodObjects)
-}
-
-function fillTable(foods) {
-  foods.reverse().forEach(function(foodType){
-    event.preventDefault
-      let button = `<td><button type='button' class='deleteButton' id='${foodType.id}'>Delete</button></td>`
-      let insertName   = `<tr id=${foodType.id}><td class="food" name="name" contentEditable>${foodType.name}</td>`
-      let insertCals   = `<td class="food" name="calories" contentEditable>${foodType.calories}`
-      let insertRow    = `${insertName}${insertCals}</td>${button}</tr>`
-      $(".foodsTable").append(insertRow)
-  })
-}
-
-function createFoodButtonListener() {
-  $('#createButton').on('click', function() {
-    $('#newFoodForm').show()
-    createFoodFormListener()
-  })
-}
-
-function createFoodFormListener() {
-  $('#createFoodForm').on('submit', function(event) {
-    event.preventDefault()
-
-    let food = $(event.currentTarget.name).val()
-    let calories = $(event.currentTarget.calories).val()
-
-    errorCheck(food, calories)
-
-    let data = $.post(herokuUrl() + `api/v1/foods?name=${food}&calories=${calories}"`)
-      .then(function(data) {
-        let button = `<td><button type='button' class='deleteButton' id='${data.id}'>Delete</button></td>`
-          let toInsert = `<tr id=${data.id}><td>${food}</td><td>${calories}</td>${button}</tr>`
-        $(".foodsTable").prepend(toInsert)
-        $('#createFoodForm')[0].reset()
-      })
-  })
-}
-
-function errorCheck(food, calories) {
-  //let words = [food, calories]
-    //let toSay = {
-      //food: "food name",
-      //calories: "calorie amount"
-    //}
-
-  //words.forEach(function(input, index) {
-    //let word = ["food", "calories"]
-    //if (!`${input}`) {
-      //$(`${word[index]}-error`).remove()
-      //throw new Error ($(`#${word[index]}Field`).append(`<p id=``${word[index]}-error``>Please enter a ${toSay[word[index]]} </p>`))
-    //}
-  //})
-//}
-
-    if (!food) {
-      $('#name-error').remove()
-      throw new Error ($('#nameField').append("<p id='name-error'>Please enter a food name</p>"))
-    }
-
-    if (!calories || calories < 1) {
-      $('#calorie-error').remove()
-      throw new Error($('#calorieField').append("<p id='calorie-error'>Please enter a calorie amount</p>"))
-    }
-}
-
-getFoods()
-
-
-module.exports = function() {
-  getFoods: getFoods
-}
+module.exports = DefaultFoodLoader
