+++ lib/index.js
@@ -1,6 +1,4 @@
 // This is the base url
 // make requests with `herokuUrl + "api/v1/foods"` etc
 const herokuUrl = require("./heroku-url")
-const food = require('./foods');
-
-
+const food = require('./foods')
